raspberry_pi_server/requirements.txt
paho-mqtt
flask
pyaudio
pygame
wave
threading
json
numpy